requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
ply==3.11
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
robotframework==7.3.2
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.7
robotframework-yamllibrary==0.2.8
rpds-py==0.27.1
urllib3==2.5.0
Werkzeug==3.1.3
